pyproject.toml
[project]
name = "bb-inflator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Josiah M. Caprino", email = "josiah.caprino@ecsdevlabs.com" },
]
requires-python = ">=3.11"
dependencies = ["click>=8.2.0", "rich>=14.0.0", "pyyaml", "gitpython"]

[project.scripts]
bb-inflator = "bb_inflator:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
